requirements.txt
pandas
numpy
requests
python-dotenv
yfinance
ta
plotly
streamlit